+++ extole-weave/build.gradle
@@ -25,7 +25,7 @@ dependencies {
     implementation project(':ai-weave')
     implementation project(':store')
     implementation project(':jira-client')
-    implementation project(':web-dev')
+    implementation project(':web-weave')
     
     testImplementation('org.springframework.boot:spring-boot-starter-test')
 }
