+++ settings.gradle
@@ -5,8 +5,8 @@ include 'store-app'
 include 'jira-client'
 include 'cyster-weave'
 include 'extole-weave'
+include 'web-weave'
 include 'weave-rest'
-include 'web-dev'
 include 'web-rest'
 include 'extole-tickets-rest'
 include 'weave-app'
