--- web-dev/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'web-dev'
