+++ web-rest/build.gradle
@@ -18,7 +18,7 @@ dependencies {
     implementation 'com.squareup.okhttp3:logging-interceptor:4.12.0'
 
     implementation project(':ai-weave')
-    implementation project(':web-dev')
+    implementation project(':web-weave')
 
     testImplementation('org.springframework.boot:spring-boot-starter-test')
 }
