+++ web-weave/src/main/java/com/cyster/web/weave/WebDeveloperSiteScan.java
@@ -1,4 +1,4 @@
-package com.cyster.web.developer;
+package com.cyster.web.weave;
 
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
@@ -7,5 +7,5 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 @Configuration
 @ComponentScan
 @EnableScheduling
-public class WebDeveloperScan {
+public class WebDeveloperSiteScan {
 }
