+++ web-weave/src/main/resources/web/simple/index.html
